storage problem fixed and download post

+++ controllers/post.js
@@ -1,6 +1,11 @@
 const Post = require("../models/post");
 const { validationResult } = require("express-validator");
 const { formatISO9075 } = require("date-fns");
+const pdf = require("pdf-creator-node");
+const fs = require("fs");
+const expath = require("path");
+
+const { fileDelete } = require("../utils/fileDelete");
 
 exports.createPost = (req, res, next) => {
   const { title, description } = req.body;
@@ -128,21 +133,6 @@ exports.updatePost = (req, res, next) => {
   const { postId, title, description } = req.body;
   const image = req.file;
   const errors = validationResult(req);
-
-  // if (image === undefined) {
-  //   return res.status(422).render("editPost", {
-  //     postId,
-  //     title,
-  //     isValidationFail: true,
-  //     title: "Post create ml",
-  //     errorMsg: "Image extension must be jpg,png or jpeg",
-  //     oldFormData: {
-  //       title,
-  //       description,
-  //     },
-  //   });
-  // }
-
   if (!errors.isEmpty()) {
     return res.status(422).render("editPost", {
       postId,
@@ -156,7 +146,6 @@ exports.updatePost = (req, res, next) => {
       isValidationFail: true,
     });
   }
-
   Post.findById(postId)
     .then((post) => {
       if (post.userId.toString() !== req.user._id.toString()) {
@@ -165,9 +154,9 @@ exports.updatePost = (req, res, next) => {
       post.title = title;
       post.description = description;
       if (image) {
+        fileDelete(post.imgUrl);
         post.imgUrl = image.path;
       }
-
       return post.save().then(() => {
         console.log("Post Updated");
         res.redirect("/");
@@ -180,12 +169,85 @@ exports.updatePost = (req, res, next) => {
     });
 };
 
+//Handle delete post
 exports.deletePost = (req, res, next) => {
   const { postId } = req.params;
-  Post.deleteOne({ _id: postId, userId: req.user._id })
+  Post.findById(postId)
+    .then((post) => {
+      if (!post) {
+        return res.redirect("/");
+      }
+      fileDelete(post.imgUrl);
+      return Post.deleteOne({ _id: postId, userId: req.user._id });
+    })
     .then(() => {
       console.log("Post Deleted!!");
       res.redirect("/");
+    })
+    .catch((err) => {
+      console.log(err);
+      const error = new Error("Something Went Wrong");
+      return next(error);
+    });
+};
+
+exports.savePostAsPDF = (req, res, next) => {
+  const { id } = req.params;
+  const templateUrl = `${expath.join(
+    __dirname,
+    "../views/template/template.html"
+  )}`;
+  const html = fs.readFileSync(templateUrl, "utf8");
+
+  const options = {
+    format: "A3",
+    orientation: "portrait",
+    border: "10mm",
+    header: {
+      height: "45mm",
+      contents:
+        '<h4 style="text-align: center;">PDF Download From Blog.io</h4>',
+    },
+    footer: {
+      height: "28mm",
+      contents: {
+        first: "Cover page",
+        contents:
+          '<span style="color: #444; text-align: center">@nyinyizin.mm</span>',
+      },
+    },
+  };
+
+  Post.findById(id)
+    .populate("userId", "email")
+    .lean()
+    .then((post) => {
+      const date = new Date();
+      const pdfSaveUrl = `${expath.join(
+        __dirname,
+        "../public/pdf",
+        date.getTime() + ".pdf"
+      )}`;
+      const document = {
+        html,
+        data: {
+          post,
+        },
+        path: pdfSaveUrl,
+        type: "",
+      };
+      pdf
+        .create(document, options)
+        .then((result) => {
+          console.log(result);
+          res.download(pdfSaveUrl, (err) => {
+            if (err) throw err;
+            fileDelete(pdfSaveUrl);
+          });
+        })
+        .catch((err) => {
+          console.log(err);
+        });
     })
     .catch((err) => {
       console.log(err);
